LayoutTests/imported/w3c/web-platform-tests/pointerevents/pointerevent_mouse_pointercapture_inactivate_pointer-expected.txt



Harness Error (TIMEOUT), message = null

FAIL setPointerCapture: pointer active in outer frame, set capture to inner frame assert_equals: Inner frame can not capture when pointer is activate in outer frame expected 0 but got 1
TIMEOUT setPointerCapture: pointer active in inner frame, set capture to outer frame Test timed out
